# app/backend/cmd/api/docs/swagger.yaml
definitions:
  common.ErrorResponse:
    properties:
      error:
        description: 'example: user not found'
        type: string
    type: object
...
  handlers.CreateUserRequest:
    properties:
      lastname:
        description: 'example: Ivanov'
        type: string
      login:
        description: 'example: ivan123'
        type: string
      name:
        description: 'example: Ivan'
        type: string
      password:
        description: 'example: passw0rd'
        type: string
      role:
        description: 'example: observer'
        type: string
    type: object
...
  handlers.UpdateUserRequest:
    properties:
      lastname:
        description: 'example: Ivanov'
        type: string
      name:
        description: 'example: Ivan'
        type: string
    type: object
  handlers.UserResponse:
    properties:
      id:
        type: integer
      lastname:
        type: string
      login:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
info:
  contact: {}
  description: this is documentation for buildefect API
  title: buildefect api
  version: "1.0"
paths:
...
  /api/users:
    get:
      consumes:
      - application/json
      description: Retrieve list of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.UserResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get users
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: New user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new user
      tags:
      - users
  /api/users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete user by numeric id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: Successfully deleted user with id {id}
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Delete user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Retrieve user by numeric id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "400":
          description: invalid id
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get user by id
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update user's name and/or lastname
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Update user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"